backup-v3.1/MiniPet/requirements.txt
pygame
Pillow
pywin32
